requirements/common.txt
PyJWT>=2.4.0
